Add ip addresses of the servers

+++ 2-do_deploy_web_static.py
@@ -6,6 +6,8 @@ from datetime import datetime
 from fabric.api import *
 import os.path
 
+env.hosts = ['34.202.158.152', '100.25.34.22']
+
 
 def do_pack():
     """Generate a .tgz file"""
@@ -21,10 +23,12 @@ def do_pack():
 
 def do_deploy(archive_path):
     """distributes an archive to your web servers"""
+    if os.path.exists(archive_path) is False:
+        return False
     try:
         file_name = archive_path.split("/")
         no_ext, ext = file_name[1].split(".")
-        put(archive_path, "/tmp ")
+        put(archive_path, "/tmp")
         run("mkdir -p /data/web_static/releases/{}".format(no_ext))
         run("tar -xzf /tmp/{} -C /data/web_static/releases/{}".format(file_name[1], no_ext))
         run("rm /tmp/{}".format(file_name[1]))
@@ -32,6 +36,7 @@ def do_deploy(archive_path):
         run("rm -rf /data/web_static/releases/{}/web_static".format(no_ext))
         run("rm -rf /data/web_static/current")
         run("ln -s /data/web_static/releases/{} /data/web_static/current".format(no_ext))
+        print("New version deployed!")
         return True
     except Exception as NotFound:
         return False
